chg usage of 'long' to make gcc happy
git-svn-id: http://vtp.googlecode.com/svn/trunk@2410 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/vtdata/vtDIB.h
@@ -47,7 +47,7 @@ public:
 	bool ReadJPEG(const char *fname);
 	bool WritePNG(const char *fname);
 
-	unsigned long GetPixel24(int x, int y) const;
+	unsigned int GetPixel24(int x, int y) const;
 	void GetPixel24(int x, int y, RGBi &rgb) const;
 	void GetPixel24From8bit(int x, int y, RGBi &rgb) const;
 	void SetPixel24(int x, int y, dword color);
